Twiddled with the forest
The forest originally used colliders to determine when to activate a
section. This was overkill given that the sections were arranged on a
grid, so I removed all the guff and just made the sections activate on
player grid-index. Should be a smidge quicker as well.
TODO: Put the active-tree instantiation into the editor, rather than on
startup.