Resources updates
I've scrapped the Resources folder as it was a nonsense.
OccludedMesh now uses a binary-heap for sorting its verts.
The heap had a horrible bug, so that was upsetting.
The ray-casts still consume most of the function's time, but that can't
be sorted without a major re-think of the algorithm.