Dazed and Occluded
- OccludedMesh now tracks which colliders in range are dynamic (have a
rigidbody above them in their hierarchy) and only updates when a dynamic
collider's transform changes. This could be of negative impact in busy
scenarios, but thankfully I am making a sedate little game. General
performance benefits are that no meshes are recalculated when the player
is idle.
- Removed a wadge of pointless behaviours that were sat on every single
facility-wall object, spitting and giving attitude.
- It turns out the startup lag is due to the snow particles spawning and
priming, which is a bit of a sod.
- Put another door in the test house and made the upstairs hall light
dynamic.
- The AI isn't navigating around trees and it seems that dynamic navmesh
carving is gluttonous in its CPU usage, so that needs a looking at.
Trees are just capsule colliders with guaranteed space between each
other, so maybe steering behaviours would be sufficient?